优化了时间选择的UI

+++ SK_UIKite/SK_UIKite/SK_Extend/SK_Tools/SK_3rdTools/GFCalendar/GFCalendarScrollView.h
@@ -16,6 +16,6 @@ typedef void (^DidSelectDayHandler)(NSString * startDate, NSString * endDate);
 @property (nonatomic, copy) DidSelectDayHandler didSelectDayHandler; // 日期点击回调
 
 - (void)refreshToCurrentMonth; // 刷新 calendar 回到当前日期月份
-- (void)refreshSelectDate_stratDateNum:(long)startDateNum endDateNum:(long)endDateNum;//选中时间段后刷新日历
+- (void)refreshSelectDate_stratDateNum:(long long)startDateNum endDateNum:(long long)endDateNum;//选中时间段后刷新日历
 
 @end
